backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
google-generativeai>=0.3.0
scikit-learn>=1.0.0
